Extract pure transition math from state_machine_manager into logic/
Move validate_transition_probabilities, calculate_aggregated_probabilities, analyze_simulation_results and create_visualizations verbatim into a new Streamlit-free module logic/transition_math.py so they can be unit-tested without a Streamlit runtime. state_machine_creator's inline import now points at logic.transition_math instead of state_machine_manager, breaking the creator->manager cross-import that dragged matplotlib/seaborn/run_simulation into the creator. Pure relocation, no behavior change; the (states) -> None contract is unchanged.
