+++ dmp/app/state_machine/logic/__init__.py
@@ -1,0 +1,6 @@
+"""Pure, Streamlit-free logic extracted from the state_machine UI modules.
+
+Modules in this package must not import ``streamlit``, ``seaborn``,
+``streamlit_javascript`` or ``core.*`` so the test suite can import them
+without a Streamlit runtime.
+"""
